Timer を TriggerDetector に組み込んでみた

+++ include/Timer.h
@@ -14,7 +14,7 @@ public:
      */
     static Timer* getInstance(void);
 
-    double current(void);
+    int current(void);
 
     /**
      * インスタンス生成時の time
